// src/app/modules/antifraudes/fraude/analizar-fraude/interfaces/fraude.interface.ts
export interface IFraude {
    componente: string;
    descripcion: string;
    empresa: string;
    fecha_fin_ocu: string;
    fecha_fin_re: string;
    fecha_ini_ocu: string;
    fecha_ini_re: string;
    fuente: string;
    id: string;
    impacto: string;
    riesgo: string;
    servicio: string;
    tipo_fraude: string;
    usuario: string;
  }
  